Re-enabled ability to use the "generateName" capability in services
This is needed to add ssh port forwarding, i.e. NodePorts, for each workspace. If we use name, it'll have collisions with other NodePorts so we need it to randomize names. generateName is a feature in Kubernetes, but Che needs a name so this fix preemptively generates a name using the same Kubernetes syntax.